Move all rack awareness tests into one suite
