--- systemtest/src/test/java/io/strimzi/systemtest/kafka/KafkaRackST.java
@@ -1,127 +1,0 @@
-/*
- * Copyright Strimzi authors.
- * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
- */
-package io.strimzi.systemtest.kafka;
-
-import io.fabric8.kubernetes.api.model.Affinity;
-import io.fabric8.kubernetes.api.model.Event;
-import io.fabric8.kubernetes.api.model.NodeAffinity;
-import io.fabric8.kubernetes.api.model.NodeSelectorRequirement;
-import io.fabric8.kubernetes.api.model.Pod;
-import io.fabric8.kubernetes.api.model.PodAffinityTerm;
-import io.strimzi.api.kafka.model.KafkaResources;
-import io.strimzi.systemtest.AbstractST;
-import static io.strimzi.systemtest.Constants.INTERNAL_CLIENTS_USED;
-import static io.strimzi.systemtest.Constants.RACK_AWARENESS;
-import static io.strimzi.systemtest.Constants.REGRESSION;
-import io.strimzi.systemtest.Environment;
-import io.strimzi.systemtest.annotations.ParallelNamespaceTest;
-import io.strimzi.systemtest.annotations.ParallelSuite;
-import static io.strimzi.systemtest.k8s.Events.Created;
-import static io.strimzi.systemtest.k8s.Events.Pulled;
-import static io.strimzi.systemtest.k8s.Events.Scheduled;
-import static io.strimzi.systemtest.k8s.Events.Started;
-import io.strimzi.systemtest.kafkaclients.internalClients.KafkaClients;
-import io.strimzi.systemtest.kafkaclients.internalClients.KafkaClientsBuilder;
-import static io.strimzi.systemtest.matchers.Matchers.hasAllOfReasons;
-import io.strimzi.systemtest.storage.TestStorage;
-import io.strimzi.systemtest.templates.crd.KafkaTemplates;
-import io.strimzi.systemtest.utils.StUtils;
-import io.strimzi.systemtest.utils.kubeUtils.objects.PodUtils;
-import static io.strimzi.test.k8s.KubeClusterResource.cmdKubeClient;
-import static io.strimzi.test.k8s.KubeClusterResource.kubeClient;
-import java.util.List;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-import static org.hamcrest.CoreMatchers.is;
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.hasEntry;
-import static org.junit.jupiter.api.Assumptions.assumeFalse;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Tag;
-import org.junit.jupiter.api.extension.ExtensionContext;
-
-@Tag(REGRESSION)
-@Tag(INTERNAL_CLIENTS_USED)
-@Tag(RACK_AWARENESS)
-@ParallelSuite
-public class KafkaRackST extends AbstractST {
-    private static final Logger LOGGER = LogManager.getLogger(KafkaRackST.class);
-
-    @ParallelNamespaceTest
-    void testKafkaRackAwareness(ExtensionContext extensionContext) {
-        assumeFalse(Environment.isNamespaceRbacScope());
-
-        TestStorage storage = new TestStorage(extensionContext);
-        String namespace = storage.getNamespaceName();
-        String clusterName = storage.getClusterName();
-        String topologyKey = "kubernetes.io/hostname";
-
-        resourceManager.createResource(extensionContext, KafkaTemplates.kafkaEphemeral(clusterName, 1, 1)
-            .editSpec()
-                .editKafka()
-                    .withNewRack(topologyKey)
-                .endKafka()
-            .endSpec()
-            .build());
-
-        LOGGER.info("Kafka cluster deployed successfully");
-        String ssName = KafkaResources.kafkaStatefulSetName(clusterName);
-        String podName = PodUtils.getPodNameByPrefix(namespace, ssName);
-        Pod pod = PodUtils.getPodByName(namespace, podName);
-
-        // check that spec matches the actual pod configuration
-        Affinity specAffinity = StUtils.getStatefulSetOrStrimziPodSetAffinity(KafkaResources.kafkaStatefulSetName(clusterName));
-        NodeSelectorRequirement specNodeRequirement = specAffinity.getNodeAffinity().getRequiredDuringSchedulingIgnoredDuringExecution().getNodeSelectorTerms().get(0).getMatchExpressions().get(0);
-        NodeAffinity podAffinity = pod.getSpec().getAffinity().getNodeAffinity();
-        NodeSelectorRequirement podNodeRequirement = podAffinity.getRequiredDuringSchedulingIgnoredDuringExecution().getNodeSelectorTerms().get(0).getMatchExpressions().get(0);
-        assertThat(podNodeRequirement, is(specNodeRequirement));
-        assertThat(specNodeRequirement.getKey(), is(topologyKey));
-        assertThat(specNodeRequirement.getOperator(), is("Exists"));
-
-        PodAffinityTerm specPodAntiAffinityTerm = specAffinity.getPodAntiAffinity().getPreferredDuringSchedulingIgnoredDuringExecution().get(0).getPodAffinityTerm();
-        PodAffinityTerm podAntiAffinityTerm = pod.getSpec().getAffinity().getPodAntiAffinity().getPreferredDuringSchedulingIgnoredDuringExecution().get(0).getPodAffinityTerm();
-        assertThat(podAntiAffinityTerm, is(specPodAntiAffinityTerm));
-        assertThat(specPodAntiAffinityTerm.getTopologyKey(), is(topologyKey));
-        assertThat(specPodAntiAffinityTerm.getLabelSelector().getMatchLabels(), hasEntry("strimzi.io/cluster", clusterName));
-        assertThat(specPodAntiAffinityTerm.getLabelSelector().getMatchLabels(), hasEntry("strimzi.io/name", KafkaResources.kafkaStatefulSetName(clusterName)));
-
-        // check Kafka rack awareness configuration
-        String podNodeName = pod.getSpec().getNodeName();
-        String hostname = podNodeName.contains(".") ? podNodeName.substring(0, podNodeName.indexOf(".")) : podNodeName;
-
-        String rackIdOut = cmdKubeClient().execInPod(KafkaResources.kafkaPodName(clusterName, 0), "/bin/bash", "-c", "cat /opt/kafka/init/rack.id").out().trim();
-        String brokerRackOut = cmdKubeClient().execInPod(KafkaResources.kafkaPodName(clusterName, 0), "/bin/bash", "-c", "cat /tmp/strimzi.properties | grep broker.rack").out().trim();
-        assertThat(rackIdOut.trim(), is(hostname));
-        assertThat(brokerRackOut.contains("broker.rack=" + hostname), is(true));
-
-        // additional checks
-        String producerName = "hello-world-producer";
-        String consumerName = "hello-world-consumer";
-        String uid = kubeClient().getPodUid(KafkaResources.kafkaPodName(clusterName, 0));
-        List<Event> events = kubeClient().listEventsByResourceUid(uid);
-        assertThat(events, hasAllOfReasons(Scheduled, Pulled, Created, Started));
-
-        KafkaClients kafkaBasicClientJob = new KafkaClientsBuilder()
-            .withProducerName(producerName)
-            .withConsumerName(consumerName)
-            .withBootstrapAddress(KafkaResources.plainBootstrapAddress(clusterName))
-            .withTopicName(TOPIC_NAME)
-            .withMessageCount(MESSAGE_COUNT)
-            .withDelayMs(0)
-            .build();
-
-        resourceManager.createResource(extensionContext, kafkaBasicClientJob.producerStrimzi());
-        resourceManager.createResource(extensionContext, kafkaBasicClientJob.consumerStrimzi());
-    }
-
-    @BeforeAll
-    void setUp() {
-        clusterOperator.unInstall();
-        clusterOperator = clusterOperator
-                .defaultInstallation()
-                .createInstallation()
-                .runInstallation();
-    }
-}
