+++ systemtest/src/test/java/io/strimzi/systemtest/specific/RackAwarenessIsolatedST.java
@@ -2,15 +2,19 @@
  * Copyright Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.strimzi.systemtest.connect;
+package io.strimzi.systemtest.specific;
 
 import io.fabric8.kubernetes.api.model.Affinity;
+import io.fabric8.kubernetes.api.model.Event;
 import io.fabric8.kubernetes.api.model.NodeAffinity;
 import io.fabric8.kubernetes.api.model.NodeSelectorRequirement;
 import io.fabric8.kubernetes.api.model.Pod;
+import io.fabric8.kubernetes.api.model.PodAffinityTerm;
 import io.fabric8.kubernetes.api.model.PodStatus;
 import io.strimzi.api.kafka.model.KafkaConnect;
 import io.strimzi.api.kafka.model.KafkaConnectResources;
+import io.strimzi.api.kafka.model.KafkaMirrorMaker2Resources;
+import io.strimzi.api.kafka.model.KafkaResources;
 import io.strimzi.api.kafka.model.Rack;
 import io.strimzi.operator.common.model.Labels;
 import io.strimzi.systemtest.AbstractST;
@@ -20,11 +24,17 @@ import static io.strimzi.systemtest.Constants.CONNECT;
 import static io.strimzi.systemtest.Constants.CONNECT_COMPONENTS;
 import static io.strimzi.systemtest.Constants.CO_OPERATION_TIMEOUT_SHORT;
 import static io.strimzi.systemtest.Constants.INTERNAL_CLIENTS_USED;
-import static io.strimzi.systemtest.Constants.RACK_AWARENESS;
 import static io.strimzi.systemtest.Constants.REGRESSION;
 import io.strimzi.systemtest.Environment;
 import io.strimzi.systemtest.annotations.IsolatedSuite;
 import io.strimzi.systemtest.annotations.IsolatedTest;
+import static io.strimzi.systemtest.k8s.Events.Created;
+import static io.strimzi.systemtest.k8s.Events.Pulled;
+import static io.strimzi.systemtest.k8s.Events.Scheduled;
+import static io.strimzi.systemtest.k8s.Events.Started;
+import io.strimzi.systemtest.kafkaclients.internalClients.KafkaClients;
+import io.strimzi.systemtest.kafkaclients.internalClients.KafkaClientsBuilder;
+import static io.strimzi.systemtest.matchers.Matchers.hasAllOfReasons;
 import io.strimzi.systemtest.resources.ResourceManager;
 import static io.strimzi.systemtest.resources.ResourceManager.cmdKubeClient;
 import static io.strimzi.systemtest.resources.ResourceManager.kubeClient;
@@ -34,8 +44,10 @@ import io.strimzi.systemtest.resources.operator.SetupClusterOperator;
 import io.strimzi.systemtest.storage.TestStorage;
 import io.strimzi.systemtest.templates.crd.KafkaClientsTemplates;
 import io.strimzi.systemtest.templates.crd.KafkaConnectTemplates;
+import io.strimzi.systemtest.templates.crd.KafkaMirrorMaker2Templates;
 import io.strimzi.systemtest.templates.crd.KafkaTemplates;
 import io.strimzi.systemtest.templates.crd.KafkaTopicTemplates;
+import io.strimzi.systemtest.utils.StUtils;
 import io.strimzi.systemtest.utils.kafkaUtils.KafkaConnectUtils;
 import io.strimzi.systemtest.utils.kafkaUtils.KafkaTopicUtils;
 import io.strimzi.systemtest.utils.kubeUtils.controllers.DeploymentUtils;
@@ -47,6 +59,7 @@ import java.util.Map;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.hasEntry;
 import static org.hamcrest.Matchers.is;
 import static org.junit.jupiter.api.Assumptions.assumeFalse;
 import static org.junit.jupiter.api.Assumptions.assumeTrue;
@@ -58,10 +71,76 @@ import org.junit.jupiter.api.extension.ExtensionContext;
 @Tag(CONNECT)
 @Tag(CONNECT_COMPONENTS)
 @Tag(INTERNAL_CLIENTS_USED)
-@Tag(RACK_AWARENESS)
 @IsolatedSuite
-class ConnectRackIsolatedST extends AbstractST {
-    private static final Logger LOGGER = LogManager.getLogger(ConnectRackIsolatedST.class);
+class RackAwarenessIsolatedST extends AbstractST {
+    private static final Logger LOGGER = LogManager.getLogger(RackAwarenessIsolatedST.class);
+
+    @IsolatedTest
+    void testKafkaRackAwareness(ExtensionContext extensionContext) {
+        assumeFalse(Environment.isNamespaceRbacScope());
+
+        TestStorage storage = new TestStorage(extensionContext);
+        String namespace = storage.getNamespaceName();
+        String clusterName = storage.getClusterName();
+        String topologyKey = "kubernetes.io/hostname";
+
+        resourceManager.createResource(extensionContext, KafkaTemplates.kafkaEphemeral(clusterName, 1, 1)
+                .editSpec()
+                    .editKafka()
+                        .withNewRack(topologyKey)
+                    .endKafka()
+                .endSpec()
+                .build());
+
+        LOGGER.info("Kafka cluster deployed successfully");
+        String ssName = KafkaResources.kafkaStatefulSetName(clusterName);
+        String podName = PodUtils.getPodNameByPrefix(namespace, ssName);
+        Pod pod = PodUtils.getPodByName(namespace, podName);
+
+        // check that spec matches the actual pod configuration
+        Affinity specAffinity = StUtils.getStatefulSetOrStrimziPodSetAffinity(KafkaResources.kafkaStatefulSetName(clusterName));
+        NodeSelectorRequirement specNodeRequirement = specAffinity.getNodeAffinity().getRequiredDuringSchedulingIgnoredDuringExecution().getNodeSelectorTerms().get(0).getMatchExpressions().get(0);
+        NodeAffinity podAffinity = pod.getSpec().getAffinity().getNodeAffinity();
+        NodeSelectorRequirement podNodeRequirement = podAffinity.getRequiredDuringSchedulingIgnoredDuringExecution().getNodeSelectorTerms().get(0).getMatchExpressions().get(0);
+        assertThat(podNodeRequirement, is(specNodeRequirement));
+        assertThat(specNodeRequirement.getKey(), is(topologyKey));
+        assertThat(specNodeRequirement.getOperator(), is("Exists"));
+
+        PodAffinityTerm specPodAntiAffinityTerm = specAffinity.getPodAntiAffinity().getPreferredDuringSchedulingIgnoredDuringExecution().get(0).getPodAffinityTerm();
+        PodAffinityTerm podAntiAffinityTerm = pod.getSpec().getAffinity().getPodAntiAffinity().getPreferredDuringSchedulingIgnoredDuringExecution().get(0).getPodAffinityTerm();
+        assertThat(podAntiAffinityTerm, is(specPodAntiAffinityTerm));
+        assertThat(specPodAntiAffinityTerm.getTopologyKey(), is(topologyKey));
+        assertThat(specPodAntiAffinityTerm.getLabelSelector().getMatchLabels(), hasEntry("strimzi.io/cluster", clusterName));
+        assertThat(specPodAntiAffinityTerm.getLabelSelector().getMatchLabels(), hasEntry("strimzi.io/name", KafkaResources.kafkaStatefulSetName(clusterName)));
+
+        // check Kafka rack awareness configuration
+        String podNodeName = pod.getSpec().getNodeName();
+        String hostname = podNodeName.contains(".") ? podNodeName.substring(0, podNodeName.indexOf(".")) : podNodeName;
+
+        String rackIdOut = cmdKubeClient().execInPod(KafkaResources.kafkaPodName(clusterName, 0), "/bin/bash", "-c", "cat /opt/kafka/init/rack.id").out().trim();
+        String brokerRackOut = cmdKubeClient().execInPod(KafkaResources.kafkaPodName(clusterName, 0), "/bin/bash", "-c", "cat /tmp/strimzi.properties | grep broker.rack").out().trim();
+        assertThat(rackIdOut.trim(), is(hostname));
+        assertThat(brokerRackOut.contains("broker.rack=" + hostname), is(true));
+
+        // additional checks
+        String producerName = "hello-world-producer";
+        String consumerName = "hello-world-consumer";
+        String uid = kubeClient().getPodUid(KafkaResources.kafkaPodName(clusterName, 0));
+        List<Event> events = kubeClient().listEventsByResourceUid(uid);
+        assertThat(events, hasAllOfReasons(Scheduled, Pulled, Created, Started));
+
+        KafkaClients kafkaBasicClientJob = new KafkaClientsBuilder()
+                .withProducerName(producerName)
+                .withConsumerName(consumerName)
+                .withBootstrapAddress(KafkaResources.plainBootstrapAddress(clusterName))
+                .withTopicName(TOPIC_NAME)
+                .withMessageCount(MESSAGE_COUNT)
+                .withDelayMs(0)
+                .build();
+
+        resourceManager.createResource(extensionContext, kafkaBasicClientJob.producerStrimzi());
+        resourceManager.createResource(extensionContext, kafkaBasicClientJob.consumerStrimzi());
+    }
 
     @IsolatedTest
     void testConnectRackAwareness(ExtensionContext extensionContext) {
@@ -285,6 +364,50 @@ class ConnectRackIsolatedST extends AbstractST {
                 .runBundleInstallation();
 
         DeploymentUtils.waitTillDepHasRolled(clusterOperator.getDeploymentNamespace(), ResourceManager.getCoDeploymentName(), 1, coSnapshot);
+    }
+
+    @IsolatedTest
+    void testMirrorMaker2RackAwareness(ExtensionContext extensionContext) {
+        assumeFalse(Environment.isNamespaceRbacScope());
+
+        TestStorage storage = new TestStorage(extensionContext);
+        String namespace = storage.getNamespaceName();
+        String clusterName = storage.getClusterName();
+        String kafkaClusterSourceName = clusterName + "-source";
+        String kafkaClusterTargetName = clusterName + "-target";
+        String topologyKey = "kubernetes.io/hostname";
+
+        resourceManager.createResource(extensionContext,
+                KafkaTemplates.kafkaEphemeral(kafkaClusterSourceName, 1, 1).build());
+        resourceManager.createResource(extensionContext,
+                KafkaTemplates.kafkaEphemeral(kafkaClusterTargetName, 1, 1).build());
+
+        resourceManager.createResource(extensionContext,
+                KafkaMirrorMaker2Templates.kafkaMirrorMaker2(clusterName, kafkaClusterTargetName, kafkaClusterSourceName, 1, false)
+                        .editSpec()
+                            .withNewRack(topologyKey)
+                        .endSpec()
+                        .build());
+
+        LOGGER.info("MirrorMaker2 cluster deployed successfully");
+        String deployName = KafkaMirrorMaker2Resources.deploymentName(clusterName);
+        String podName = PodUtils.getPodNameByPrefix(namespace, deployName);
+        Pod pod = PodUtils.getPodByName(namespace, podName);
+
+        // check that spec matches the actual pod configuration
+        Affinity specAffinity = kubeClient().getDeployment(deployName).getSpec().getTemplate().getSpec().getAffinity();
+        NodeSelectorRequirement specNodeRequirement = specAffinity.getNodeAffinity().getRequiredDuringSchedulingIgnoredDuringExecution().getNodeSelectorTerms().get(0).getMatchExpressions().get(0);
+        NodeAffinity podAffinity = pod.getSpec().getAffinity().getNodeAffinity();
+        NodeSelectorRequirement podNodeRequirement = podAffinity.getRequiredDuringSchedulingIgnoredDuringExecution().getNodeSelectorTerms().get(0).getMatchExpressions().get(0);
+        assertThat(podNodeRequirement, is(specNodeRequirement));
+        assertThat(podNodeRequirement.getKey(), is(topologyKey));
+        assertThat(podNodeRequirement.getOperator(), is("Exists"));
+
+        // check Kafka client rack awareness configuration
+        String podNodeName = pod.getSpec().getNodeName();
+        String hostname = podNodeName.contains(".") ? podNodeName.substring(0, podNodeName.indexOf(".")) : podNodeName;
+        String commandOut = cmdKubeClient().execInPod(podName, "/bin/bash", "-c", "cat /tmp/strimzi-connect.properties | grep consumer.client.rack").out().trim();
+        assertThat(commandOut.equals("consumer.client.rack=" + hostname), is(true));
     }
 
     @BeforeAll
